refact/front/PassengerDriverBtn.jsx: moves to src/components/buttons/PassengerDriverBtn.jsx
